requirements.txt
aiogram==3.13.1
PyMySQL==1.1.1
python-dotenv==1.0.1
